requirements.txt
pyYAML
torch
transformers
sentence-transformers
easyocr
chromadb
Pillow